中規模ベンチマークの動作検証確認

+++ src/CacheSim.hpp
@@ -14,7 +14,7 @@ class CacheSim{
     private:
         const std::string fname;
         CacheMemory *cl1, *cl2, *cl3;
-        int access, l1miss, l2miss, l3miss, l1hits, l2hits, l3hits;
+        long long int access, l1miss, l2miss, l3miss, l1hits, l2hits, l3hits;
         std::list<uint64_t> addrList;
         bool isFullyAssociative;
     public:
